--- a/source/slang/slang-ast-support-types.h
+++ b/source/slang/slang-ast-support-types.h
@@ -551,13 +551,11 @@ namespace Slang
         SLANG_VALUE_CLASS(QualType) 
 
         Type*	type = nullptr;
-        bool	        isLeftValue;
+        bool	        isLeftValue = false;
         bool            hasReadOnlyOnTarget = false;
         bool	        isWriteOnly = false;
 
-        QualType()
-            : isLeftValue(false)
-        {}
+        QualType() = default;
 
         QualType(Type* type);