… source/slang/slang-check-inheritance.cpp
… source/slang/slang-emit.cpp
… source/slang/slang-ir-autodiff-fwd.cpp
… source/slang/slang-ir-ssa-simplification.cpp
… source/slang/slang-options.cpp
… source/slangc/main.cpp
… tests/bindings/multiple-parameter-blocks.slang
… tests/language-feature/extensions/extension-method-simple.slang