1. Add caching for implicit conversion overload resolution results.
2. Handle invalidation of type checking cache upon extension registration more gracefully.
3. Add a setting to `simplifyIR` to disable `redundancyRemoval` pass (still enabled by default).