@jkwak-work 2024-04-02
I have some timing numbers here.

Without this change, slangc takes 47.19s ~ 48.18s for compiling 40 TSR shaders.
With this change, slangc takes slightly less time, 46.19s ~ 47.71s for the same TSR shaders.

Without this change, FXC took 90.28s ~ 91.69s for compiling 6 TSR shaders.
With this change, FXC took little longer, 94.42s ~ 95.10s for the same TSR shaders.

I noticed that with this changes, the number of HLSL lines generated from Slangc is about 3% less.
I am not sure if it is expected.