@csyonghe 2024-04-02
Given your measurements, I am not sure if we should be merging these changes, since it doesn't seem to be a net win.