@csyonghe 2024-05-29
Update: the changes here resulted 20% speedup in frontend when processing typical shader workloads, so I am going to merge it.